requirements.txt
Flask
Werkzeug
SQLAlchemy
PyPDF2
python-docx
requests